Changed type of TrafficObject.id to make it consistent with Junction.

+++ scenario_ref/include/model/TrafficObject.h
@@ -25,12 +25,12 @@ public:
     };
 
 
-    TrafficObject(int id, double length=0, double x=0, double v=0, double a=0) : id(id), x(x),v(v), a(a), length(length), lane(nullptr) {};
+    TrafficObject(uint64_t id, double length=0, double x=0, double v=0, double a=0) : id(id), x(x),v(v), a(a), length(length), lane(nullptr) {};
 
     /**
      * state. put acceleration in here for a more generic implementation of Car::nextStep
      */
-    int id;
+    uint64_t id;
     double v;
     double a;
     double length;
